# net/art: allow non-pointers as values