Values are still turned into pointers internally to maintain the invariants of strideTable, but from the user's perspective it's now possible to tbl.Insert(pfx, true) rather than
tbl.Insert(pfx, ptr.To(true)).

Updates #7781

---

The benchmarks are a little noisy, but generally positive: insertion and deletion got 10-50% faster, and the route tables are half the size they were before. We do pay 1 extra alloc per insertion now (~3 allocs/insert instead of ~2), but given the prior API that's just an alloc from outside the table that's now moved inside, for most uses. Full Table benchmarks below.

```
goos: linux
goarch: amd64
pkg: tailscale.com/net/art
cpu: AMD Ryzen 9 5950X 16-Core Processor
                          │ before-tbl.txt │            after-tbl.txt             │
                          │     sec/op     │    sec/op     vs base                │
TableInsertion/ipv4/10         809.1n ± 1%   516.1n ±  0%  -36.21% (p=0.000 n=10)
TableInsertion/ipv4/100        906.9n ± 1%   617.4n ±  0%  -31.92% (p=0.000 n=10)
TableInsertion/ipv4/1000      1601.0n ± 3%   996.8n ±  2%  -37.74% (p=0.000 n=10)
TableInsertion/ipv4/10000      2.068µ ± 4%   1.378µ ±  5%  -33.39% (p=0.000 n=10)
TableInsertion/ipv6/10        1169.5n ± 1%   641.2n ±  0%  -45.17% (p=0.000 n=10)
TableInsertion/ipv6/100       1453.0n ± 7%   747.2n ±  0%  -48.58% (p=0.000 n=10)
TableInsertion/ipv6/1000       2.307µ ± 3%   1.401µ ±  2%  -39.26% (p=0.000 n=10)
TableInsertion/ipv6/10000      2.945µ ± 4%   1.930µ ±  3%  -34.47% (p=0.000 n=10)
TableDelete/ipv4/10            358.7n ± 2%   285.8n ±  1%  -20.34% (p=0.000 n=10)
TableDelete/ipv4/100           284.4n ± 1%   265.8n ±  3%   -6.54% (p=0.000 n=10)
TableDelete/ipv4/1000          406.2n ± 2%   363.9n ±  1%  -10.44% (p=0.000 n=10)
TableDelete/ipv4/10000         625.2n ± 7%   460.3n ±  6%  -26.38% (p=0.000 n=10)
TableDelete/ipv6/10            347.5n ± 1%   352.6n ±  1%   +1.50% (p=0.004 n=10)
TableDelete/ipv6/100           502.1n ± 0%   250.3n ±  0%  -50.15% (p=0.000 n=10)
TableDelete/ipv6/1000          516.4n ± 3%   425.3n ±  1%  -17.65% (p=0.000 n=10)
TableDelete/ipv6/10000         702.4n ± 4%   612.4n ± 10%  -12.82% (p=0.000 n=10)
TableGet/ipv4/10               48.19n ± 0%   47.40n ±  0%   -1.63% (p=0.000 n=10)
TableGet/ipv4/100              52.07n ± 0%   51.46n ±  0%   -1.18% (p=0.000 n=10)
TableGet/ipv4/1000             56.82n ± 1%   56.93n ±  1%        ~ (p=0.425 n=10)
TableGet/ipv4/10000            75.08n ± 1%   79.38n ±  1%   +5.73% (p=0.001 n=10)
TableGet/ipv6/10               47.08n ± 1%   38.83n ±  0%  -17.51% (p=0.000 n=10)
TableGet/ipv6/100              55.43n ± 0%   43.31n ±  2%  -21.87% (p=0.000 n=10)
TableGet/ipv6/1000             58.41n ± 2%   57.67n ±  1%   -1.28% (p=0.029 n=10)
TableGet/ipv6/10000            76.19n ± 4%   78.80n ±  4%        ~ (p=0.063 n=10)
geomean                        340.1n        264.7n        -22.17%

                          │ before-tbl.txt │             after-tbl.txt              │
                          │    avg-B/op    │   avg-B/op    vs base                  │
TableInsertion/ipv4/10       11.941Ki ± 0%   4.528Ki ± 0%  -62.08% (p=0.000 n=10)
TableInsertion/ipv4/100      10.870Ki ± 0%   5.758Ki ± 0%  -47.03% (p=0.000 n=10)
TableInsertion/ipv4/1000     10.405Ki ± 0%   5.109Ki ± 0%  -50.90% (p=0.000 n=10)
TableInsertion/ipv4/10000     9.099Ki ± 0%   4.356Ki ± 0%  -52.12% (p=0.000 n=10)
TableInsertion/ipv6/10       15.928Ki ± 0%   7.086Ki ± 0%  -55.51% (p=0.000 n=10)
TableInsertion/ipv6/100      13.664Ki ± 0%   7.168Ki ± 0%  -47.54% (p=0.000 n=10)
TableInsertion/ipv6/1000     14.829Ki ± 0%   7.151Ki ± 0%  -51.77% (p=0.000 n=10)
TableInsertion/ipv6/10000    13.220Ki ± 0%   6.333Ki ± 0%  -52.09% (p=0.000 n=10)
TableDelete/ipv4/10             3.220 ± 0%     2.400 ± 0%  -25.47% (p=0.000 n=10)
TableDelete/ipv4/100            3.940 ± 0%     3.840 ± 0%   -2.54% (p=0.000 n=10)
TableDelete/ipv4/1000           3.980 ± 0%     3.980 ± 0%        ~ (p=0.211 n=10)
TableDelete/ipv4/10000          4.000 ± 0%     4.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/10             16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/100            15.84 ± 0%     16.00 ± 0%   +1.01% (p=0.000 n=10)
TableDelete/ipv6/1000           15.98 ± 0%     15.99 ± 0%   +0.06% (p=0.000 n=10)
TableDelete/ipv6/10000          16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                         312.6          211.1       -32.47%
¹ all samples are equal

                          │ before-tbl.txt │              after-tbl.txt              │
                          │ avg-allocs/op  │ avg-allocs/op  vs base                  │
TableInsertion/ipv4/10          1.800 ± 0%      2.600 ± 0%  +44.44% (p=0.000 n=10)
TableInsertion/ipv4/100         1.800 ± 0%      2.880 ± 0%  +60.00% (p=0.000 n=10)
TableInsertion/ipv4/1000        1.780 ± 0%      2.800 ± 0%  +57.30% (p=0.000 n=10)
TableInsertion/ipv4/10000       1.690 ± 0%      2.680 ± 0%  +58.58% (p=0.000 n=10)
TableInsertion/ipv6/10          2.100 ± 0%      3.000 ± 0%  +42.86% (p=0.000 n=10)
TableInsertion/ipv6/100         2.010 ± 0%      3.110 ± 0%  +54.73% (p=0.000 n=10)
TableInsertion/ipv6/1000        2.120 ± 0%      3.120 ± 0%  +47.17% (p=0.000 n=10)
TableInsertion/ipv6/10000       2.000 ± 0%      2.990 ± 0%  +49.50% (p=0.000 n=10)
TableDelete/ipv4/10           1000.0m ± 0%     900.0m ± 0%  -10.00% (p=0.000 n=10)
TableDelete/ipv4/100           990.0m ± 0%     990.0m ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv4/1000           1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv4/10000          1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/10             1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/100           990.0m ± 0%    1000.0m ± 0%   +1.01% (p=0.000 n=10)
TableDelete/ipv6/1000           1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/10000          1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                         1.379           1.688       +22.43%
¹ all samples are equal

                          │ before-tbl.txt │             after-tbl.txt             │
                          │    routes/s    │   routes/s    vs base                 │
TableInsertion/ipv4/10         1.236M ± 1%    1.937M ± 0%   +56.74% (p=0.000 n=10)
TableInsertion/ipv4/100        1.103M ± 1%    1.620M ± 0%   +46.90% (p=0.000 n=10)
TableInsertion/ipv4/1000       624.7k ± 3%   1003.3k ± 2%   +60.60% (p=0.000 n=10)
TableInsertion/ipv4/10000      483.5k ± 4%    726.0k ± 5%   +50.16% (p=0.000 n=10)
TableInsertion/ipv6/10         855.0k ± 1%   1559.4k ± 0%   +82.39% (p=0.000 n=10)
TableInsertion/ipv6/100        690.7k ± 7%   1338.4k ± 0%   +93.76% (p=0.000 n=10)
TableInsertion/ipv6/1000       433.5k ± 3%    713.7k ± 2%   +64.63% (p=0.000 n=10)
TableInsertion/ipv6/10000      339.5k ± 4%    518.2k ± 3%   +52.62% (p=0.000 n=10)
TableDelete/ipv4/10            2.788M ± 2%    3.499M ± 1%   +25.53% (p=0.000 n=10)
TableDelete/ipv4/100           3.516M ± 1%    3.762M ± 3%    +7.01% (p=0.000 n=10)
TableDelete/ipv4/1000          2.461M ± 2%    2.748M ± 1%   +11.66% (p=0.000 n=10)
TableDelete/ipv4/10000         1.599M ± 8%    2.172M ± 6%   +35.82% (p=0.000 n=10)
TableDelete/ipv6/10            2.878M ± 1%    2.836M ± 1%    -1.48% (p=0.004 n=10)
TableDelete/ipv6/100           1.992M ± 0%    3.996M ± 0%  +100.63% (p=0.000 n=10)
TableDelete/ipv6/1000          1.936M ± 3%    2.351M ± 1%   +21.44% (p=0.000 n=10)
TableDelete/ipv6/10000         1.424M ± 4%    1.633M ± 9%   +14.71% (p=0.000 n=10)
geomean                        1.212M         1.723M        +42.11%

                    │ before-tbl.txt │            after-tbl.txt            │
                    │    addrs/s     │   addrs/s    vs base                │
TableGet/ipv4/10         20.75M ± 0%   21.10M ± 0%   +1.66% (p=0.000 n=10)
TableGet/ipv4/100        19.20M ± 0%   19.43M ± 0%   +1.20% (p=0.000 n=10)
TableGet/ipv4/1000       17.60M ± 1%   17.56M ± 1%        ~ (p=0.436 n=10)
TableGet/ipv4/10000      13.32M ± 1%   12.60M ± 1%   -5.42% (p=0.002 n=10)
TableGet/ipv6/10         21.24M ± 1%   25.75M ± 0%  +21.23% (p=0.000 n=10)
TableGet/ipv6/100        18.04M ± 0%   23.09M ± 2%  +28.00% (p=0.000 n=10)
TableGet/ipv6/1000       17.12M ± 1%   17.34M ± 1%   +1.29% (p=0.029 n=10)
TableGet/ipv6/10000      13.13M ± 4%   12.69M ± 4%        ~ (p=0.063 n=10)
geomean                  17.31M        18.17M        +4.99%

                    │ before-tbl.txt │            after-tbl.txt            │
                    │      B/op      │    B/op     vs base                 │
TableGet/ipv4/10        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/100       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/1000      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/10000     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/100       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/1000      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10000     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                    │ before-tbl.txt │            after-tbl.txt            │
                    │   allocs/op    │ allocs/op   vs base                 │
TableGet/ipv4/10        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/100       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/1000      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/10000     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/100       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/1000      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10000     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```